requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
bcrypt==4.0.1
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
Faker==20.1.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
looseversion==1.3.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
redis==5.0.1
rethinkdb==2.4.10
rq==1.15.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
